requirements.txt
beautifulsoup4==4.11.2
discord.py==2.1.0
requests==2.27.1
translate==3.6.1
yt_dlp==2023.1.6
